requirements.txt
annotated-types==0.7.0
astroid==3.3.5
black==24.8.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dill==0.3.9
docstring_parser==0.16
feedparser==6.0.11
ffmpeg-python==0.2.0
future==1.0.0
google-api-core==2.24.2
google-auth==2.35.0
google-cloud-aiplatform==1.70.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-speech==2.27.0
google-cloud-storage==2.18.2
google-cloud-translate==3.16.0
google-cloud-videointelligence==2.13.5
google-cloud-vision==3.7.4
google-genai==1.16.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.67.0
idna==3.10
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyink==24.10.0
pylint==3.3.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
sgmllib3k==1.0.0
shapely==2.0.6
six==1.16.0
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
vertexai==1.70.0
gunicorn==23.0.0
flask==3.1.1
